+++ src/Cru/Constants.h
@@ -201,6 +201,7 @@ static constexpr uint32_t DATA_MIDTRG(0x2);*/
 /// Registers for getting TTC info
 static constexpr Register TTC_ONU_STICKY(0x00200014);
 static constexpr Register TTC_ONU_STICKY_MON(0x00200014);
+static constexpr Register TTC_ONU_GLITCH_COUNTER(0x00200020);
 static constexpr Register TTC_PON_QUALITY(0x0010000C);
 
 /// Registers used for TTC calibration