+++ doc/releaseNotes.md
@@ -44,5 +44,7 @@ This file describes the main feature changes for released versions of ReadoutCar
 ## v0.42.3 - 28/04/2023
 - Fix for roc-status --monitoring: glitchCounter unsigned.
 
-## next version
-- Added some (per link) counters for roc-status --monitoring: pktProcessed, pktErrorProtocol, pktErrorCheck1, pktErrorCheck2, pktErrorOversize, orbitSor
+## v0.43.0 - 24/05/2023
+- Added some counters for roc-status:
+  - link (with --monitoring option only): pktProcessed, pktErrorProtocol, pktErrorCheck1, pktErrorCheck2, pktErrorOversize, orbitSor (including for UL link 15)
+  - onu (with --onu option, all output modes, including --monitoring): glitchCounter